// src/app/shared/types/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.12 (cd3cf9e)"
  }
  public: {
    Tables: {
// ...
      users: {
        Row: {
          email: string
          id: string
          phone_number: number
          username: string | null
        }
        Insert: {
          email: string
          id: string
          phone_number: number
          username?: string | null
        }
        Update: {
          email?: string
          id?: string
          phone_number?: number
          username?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      generate_gpu_benchmark_avgs: {
        Args: Record<PropertyKey, never>
        Returns: undefined
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}